Add ValidationErrorBlueprint & update OpenAPI spec
This adds a ValidationErrorBlueprint that will be used in the POST answer
feedback endpoint when an answer already has had feedback provided for it (422).

It sets the message to 'Unproccessable entity' as the default message.

If needed we will extend this to take a message and use
'Unproccessable entity' as default down the line.

Co-authored-by: Sunny Shafiq <sunny.shafiq@digital.cabinet-office.gov.uk>