Add blueprints for conversation and question
We want to render a conversation in the API, so we need a blueprint for
that. The conversation response will include the conversation's questions
too, so we need a blueprint to render those.

We pass in the answered/pending questions as options rather than doing
the query in the blueprint as we think it makes sense to do the query in
the controller and pass records in as options rather than doing any
queries in the blueprints. It feels like AR queries are more the domain
of the controller.