Validate requests with Committee middleware
Committee provides schema validation for requests. When invalid it returns
a 400 with an error message outlining the issue with the request body.

Here's the example response shown in the Committee documentation:

{
  "id":"invalid_response",
  "message":"Missing keys in response: archived_at, owner:email, owner:id"
}

There is an option to build your own validation errors
https://github.com/interagent/committee?tab=readme-ov-file#validation-errors.

This commit adds the Committee middleware to the API, and configures it
to use the OpenAPI schema.

It also adds a custom error class for request which simply strips the id
from the JSON in the response body.

It also adds a test to ensure that the middleware is working as expected
for the POST feedback endpoint.