Allow docs/api_openapi_specification.yml in Docker
The openapi specification is a runtime dependency so needs to be
included in a docker build. I've taken the step of just allowing that
one file so we don't put all the other data in, this might prove foolish
as we'll hit another annoying issue if we add another openapi file or
rename this one. Let's see.