Return pass/fail status for each output guardrail
Output guardrail results now include a hash keyed by guardrail name with
a boolean pass/fail flag, rather than an array that only contained the
names of guardrails that failed.

This is in order to provide the exact number of guardrail definitions,
as well as their names to the evaluation tool (1). Without providing
these details they must be hardcoded in the evaluation codebase,
resulting in potential discrepancies if guardrail definitions are
updated.

1: alphagov/govuk-chat-evaluation-prototype#24