Return jailbreak guardrail errors in evaluation task
This is an approach to return the data for a jailbreak guardrail that
has been returned misformatted.

The motivation for making this change is to assist with the task of
optimising the prompt, where it is currently breaking when Claude
returns an unexpected response.

The expectation is that a client of this rake task will check whether
the JSON response has a success or response_error key to decide what to
do further with the evaluation.

We may well want to use this same approach on other LLM components.