Update QuestionsFilter to include SignonUser filter
This pretty much copy and pastes the existing code for EarlyAccessUser.
If an id is passed in it uses that to modify the scope to only include
conversations associated with that signon user.

I've chosen not to filter by signon user if an early access user id is
passed in via user_id, as both filters are mutually exclusive.