Update SignonUser factory to provide name
I think name is actually going to be more informative to us here provided
the user has given it a reasonable name on creation of the API user on
Signon. If we use the email it will likely end up being a devs email
address which is less useful than something like "GOV.UK App team API user".