Add namespaced OpenAI text embedding class
We want to start experimenting with different embedding models to see
if we can improve retrieval. To do so, we'll be implementing multiple
providers in this codebase concurrently.

To make things easier when we add some others, this commit refactors the
existing `TextToEmbedding` class. We'll have a simple interface in this
class which switches on the given provider and calls the
provider-specific class.

Currently we only support OpenAI, and this is given as a default
argument. In a later commit we'll remove this default argument and use
an environment variable instead.