Allow embedding provider to be configured in retrieval
Lets us configure which provider we'll use when fetching search results
via an environment variable. Currently only OpenAI is supported.

We're not configuring the provider for embedding when indexing GOV.UK
content yet, as we'll need to embed using all providers so that the
embed content is in the index.