Remove default llm_provider arg from text embedding
This default argument was added to make the refactor a little easier,
but we want to be explicit when we call for text to be converted to an
embedding.

The reason being that we use this class for both indexing content and
retrieving documents. In each of these scenarios we might want to use
different providers, e.g. we might want to index with both OpenAI and
Claude, but only retrieve with Claude.