Allow OpenSearch embedding field name to be configurable
In order to support multiple embedding providers (for evaluating
different ones), we'll need to store these different embeddings in
different fields in OpenSearch.

So we'll need to be able to specify the embedding provider we're using
and use that to determine the field name.

Currently we're just supporting OpenAI, with a view to extending this
soon.