Setup IP based rate limiting for API endpoints
This builds on the exisiting work done by Kev to add rate limiting
to the API endpoints. The limits i've picked are somewhat arbitrary.

After a bit of reasearch on, it looks like the drawbacks of using IP
based rate limiting will be exacerbated by the fact it'll be used
on mobile devices.

Primarly due to NAT (Network Address Translation) is a common practice in mobile
networks, where multiple devices share a single public IP address.

Due to that if we're going to rate limit by IP, we will need to set quite
a high limit to avoid blocking users who are behind the same NAT.

I've somewhat arbitrarily set the limit to 1000 GET requests per minute
and 150 POST requests per minute to drive discussion. The reason the GET
requests are so high is that expect the app team to poll our answer
endpoint around once a second, which would be 60 requests per minute.

POSTs and PUTs are limited lower as they are more expensive. I've been
generous with the limit for now, but we could probably reduce that in
future.

We will also need to consider that down the line if we move the API behind
Fastly and/or use load balancing, then we will need to pass the users
IP address to us via headers.

Overall i think us rate limiting by IP is probably not a great idea, but
i though it was a good first step.

I'll add a couple of commits with our approaches for us to discuss.