Throttle API endpoints by token
We want to implement 2 different throttling strategies for API endpoints:

1. Throttle by token: This will limit the number of requests a single token
can make to a specific endpoint within a given time and will protect us
being spammed by a single products API consumption
2. Throttle by device id: This will limit the number of requests a single
device id can make to a specific endpoint within a given time and will
protect us from a single device id spamming the API

This implements the first strategy. I've used abritrary limits for now
as it's something we will need to discuss with the App team. For now the
limits are:
- 10000 read requests per minute for GET/HEAD/OPTIONS requests.
- 1000  writes requests per munute for POST/PUT/PATCH/DELETE requests.