Throttle by token
Another strategy we could use it to throttle by bearer token. This would
allow us to limit the number of requests the app could make in a given
time. The pros are:

- It would be easy to implement
- We don't need any additional info passed in the headers to identify
individual users
- It would help stop our endpoints getting spammed by users with no tokens since
all requests with no token in the header would be rejected once the limit
is reached (provided we set a default value like i've done)

The cons are:
- It would be easy to bypass by simply using a different fake token in the
header
- We don't get to identify individual users, so we're not able to use
a granular approach to rate limiting
- We would need to ensure we've got a high enough limit to allow for peaks
in usage, but not so high that it allows for abuse

We coukd mitigate this by rate limting by token and IP if we'd like. We
could se the IP to a value high enough to allow for peaks in usage,
but not so high that it allows for abuse.

I think it's probs overkill for now though.