Throttle by device id
Another option we could consider is to throttle by device id. We would
need the app team to pass us this in the headers, but there are some
advantages to this approach.

The pros are:

- It would allow us to throttle by device id, which is a more stable
  identifier than IP address (vpns, shared wifi networks etc.)
- It would allow us to throttle per user, which allows us to set lower
limits, giving malicious users far less requests before they are
  throttled.

The cons are:

- the app team would need to implement this on their side for us to
be able to use it
- requests without a device id in the headers wouldn't be throttled so if
something went wrong with the app teams implementation, we could end up
not rate limiting requests at all

We could mitigate this by also having a throttle by token which is higher
than we would expect their users to hit, but would catch malicious behaviour.

I've gone with 120 GETs per minute since we need to handle our answer endpoint
being polled by the app team. Since they should be polling once per second
i've doubled that to be safe.

20 POST/PUTs per minute is also generous. I'd expect indiviual users not
to get anywhere close to that since answers take ~7-10 seconds to generate
so if they ask a question, then give feedback then type another one i'd
be surprised if they hit anywhere close to 10.