Throttle by device id
We also want to be able to throttle by device id, so that we can
limit the number of requests from a single device using the API. I've set
reasonably generous limits for this too, but we can adjust them later if
needed.

We will need to ensure that the app team passes the device id the headers.