Add class for embedding with Titan
Adds a `TextToEmbedding::Titan` class that calls out to Titan via
Bedrock to embed a string or array of strings of text.

Works in exactly the same way that the current OpenAI one does, and the
wrapper `TextToEmbedding` class will call this class if the env var is
set to use Titan as the embedding provider.