Allow OpenSearch to be queried by Titan embedding
This extends the `ChunkedContentRepository` class to reflect OpenSearch
instances where we have a `titan_embedding` field containing the content
of a document embedded by Titan.

It follows the logic we currently have for OpenAI in that we have an
`titan_embedding` field in each OpenSearch document and the
`ChunkedContentRepository#search_by_embedding` method takes the provider
as an argument and queries against the relevant field.