Allow OpenSearch to be queried by Titan embedding
Now that we have a way of embedding text via Titan, we want a way of
querying OpenSearch by that embedding.

This follows the logic we currently have for OpenAI in that we have an
`titan_embedding` field in each OpenSearch document and the
`ChunkedContentRepository#search_by_embedding` method takes the provider
as an argument and queries against the relevant field.