Generate Titan embeddings for documents from queue
When we pull content items off the GOV.UK message queue, we want to
generate the Titan embeddings (as well as the OpenAI ones as we do
currently).

This will allow us to keep the embeddings up-to-date and query against
them when retrieving documents.

Note that we always want to generate both embeddings; we don't want to
rely on an env var here or anything. We'll only use the env var when
retrieving documents. But we want to be able to retrieve using both the
available embedding fields.