Raise a warning for generate_answer for error status
This change is to make it a bit easier to debug error situations with
the generate_answer command. This will raise a warning with some details
of the error so the user has a thread to pull to work out why. Example
usage:

```
➜  govuk-chat git:(warn-on-error) ✗ bundle exec rake evaluation:generate_answer[claude_structured_answer] QUESTION="what are the vat rates?"
Warning: answer has an error status: error_answer_service_error
The security token included in the request is expired
{"message":"Sorry, something went wrong while trying to answer your question. Try again later.\n\nWe saved your conversation. Check [GOV.UK guidance for businesses](https://www.gov.uk/browse/business) if you need information now.\n"}
```