Include additional data in evaluation:generate_report
This changes the data that is returned from Evaluation::ReportGenerator
to hold much richer information about the answer. It essentially includes
every field we store in the database aside from the ones created at
write time (ids and timestamps). This has been done to provide the data
scientists with a much greater insight into what was produced.

I've enhanced the retrieved_context to include all sources, not just
used ones, with a flag as to which ones are used - as per the request.
I've also modified the handling of the extreme edge case of search data
changing while the task is running to be flagged with a boolean as to
whether the search data is available.

A somewhat janky thing is that the answer data is returned in a hash
that hash string keys rather than symbols. I didn't see it being very
valuable to correct this as the expectation is that the data is dumped
to JSON and all becomes string. We may want to fix this is if we ever
have other usages for this class.

Originally in my chat with Nick we were going to make use of the
existing serialize code for export (serialize_for_export) however when I
started implementing I realised that would lease the extra search data
this task already has and leave a bunch of null values for ids and
timestamps.