Add rate limit headers to API respones
We want to return limit usage in the headers of the API responses so that
consumers can see how many requests they have left and when the limit
will reset.

While RackAttack doesn't support this out of the box, it does provide
some information in the env hash that we can use to set the headers

```
env["rack.attack.throttle_data"]

{
  "read requests to Conversations API with device id" =>
  {
    discriminator: "test-device-123",
    count: 2,
    period: 60,
    limit: 1,
    epoch_time: 1747308452
  }
}
``

This means we can simply iterate accross the throttles and set the headers
based on the data.

We're sending back the following headers when throttles are present for read
and write responses:

Read:

- Govuk-Api-User-Read-Rate-Limit
- Govuk-Api-User-Read-Rate-Limit-Remaining
- Govuk-Api-User-Read-Rate-Limit-Reset
- Govuk-Client-Device-Id-Read-Rate-Limit
- Govuk-Client-Device-Id-Read-Rate-Limit-Remaining
- Govuk-Client-Device-Id-Read-Rate-Limit-Reset

Write:
- Govuk-Api-User-Write-Rate-Limit
- Govuk-Api-User-Write-Rate-Limit-Remaining
- Govuk-Api-User-Write-Rate-Limit-Reset
- Govuk-Client-Device-Id-Write-Rate-Limit
- Govuk-Client-Device-Id-Write-Rate-Limit-Remaining
- Govuk-Client-Device-Id-Write-Rate-Limit-Reset

One thing that is of note about this is that i had to wire in the middleware
between ActionDispatch::Executor and the exception middleware so that we
can add headers while the response is mutable.