Ensure LinkTokenMapper#mapping uses absolute URLs
At the moment we're using relative URLs to store the links to our chunks
in the mappings. This means when we convert our links from tokens to the
URLs and render them on the page they're relative links and blow up
on Heroku since they vist heroku.com/relative-path rather than gov.uk/relative-path.

We do convert relative links to absolute URls in the #map_links_to_tokens method.
This is used to format links within the content of a chunk.
https://github.com/alphagov/govuk-chat/blob/960ce6dbb3fabc9aac114933b26b18febfa6ab78/lib/answer_composition/pipeline/context.rb#L79-L81

But we call LinkTokenMapper#map_link_to_token directly on the chunks url.
https://github.com/alphagov/govuk-chat/blob/960ce6dbb3fabc9aac114933b26b18febfa6ab78/lib/answer_composition/pipeline/context.rb#L75

This means that it was never converted to an absolute URL.

By moving the logic to convert relative URLs to absolute URLs into the
LinkTokenMapper#map_link_to_token method we can ensure that all links
are converted to absolute URLs.