Update context to use sources url instead of exact path
Now we're storing mapping the token to an absolute url, we need to use the
url to determine which sources were used to generate the answer.

This updates #update_sources_from_exact_paths_used to #update_sources_from_exact_urls_used
and updates the context to use the sources url instead of the exact path.