Return triggered result object for unexpected llm_guardrail_result
With Claude we don’t have the ability we had with OpenAI to specify that
a model can only return a single token from a limited list of options
(thus our ability to force a 0 or 1 response).

We therefore want to avoid a situation where the system raises an exception
when the model reaches the max tokens limit, and passes back an unexpected
value in the response for "llm_guardrail_result".

This updates the jailbreak checker to return a triggered result rather
than raising an exception.

Since we're no longer checking for fail_value i've removed all the related
code.