Update conversation history to be 90 days
A product decision has been made today by Sam Dub that for the app beta
we will extend the 30 days conversation history to be 90 days.

The simplest way to apply this is to apply it to both web and API, since
we don't have an active web version where users will be affected by
this.

A knock-on effect of this is that we will likely need to adjust our
deletion task (that is currently not used) that deletes conversations
after 3 months [1].

[1]: https://github.com/alphagov/govuk-chat/blob/0bc8a4a466f812124bb8c2c182e4d20bdb6e43eb/lib/tasks/date_retention.rake