Add web-chat permission to SignonUser model and seeds
We're adding a permission to ensure that only users with the 'web-chat'
permission can access the application via the web chat UI.

This adds the 'web-chat' permission to the SignonUser model and updates
the seeds file to include the permission for the user.

I've already added the permission on Signon in all 3 envs.

I've also added a trait for the factory.