Search results eval: return weighted score
This evaluation task returns the search results for a given input. DS
want the weighted_score to be returned and included in the results in
the Python eval codebase to help identify/investigate any issues.

It improves the spec a bit too as currently we're stubbing
`Search::ResultsForQuestion#call` to return an array of
`Search::ChunkedContentRepository::Result` objects, whereas in reality
the method returns an array of
`Search::ResultsForQuestion::WeightedResult`.