Add option for returning questions before a specific timestamp
In order to support pagination in the conversation API, we'll need a way
of getting the questions for a given conversation that were created
before a specific timestamp.

When the client requests a conversation, we'll send them the first 25
questions. The idea is that the client will make another request with
the timestamp if the oldest record in that batch. We'll take that
timestamp and pass it to this method, which returns 25 records older
than the timestamp, i.e. the next page.

We're using milliseconds here to be safe. It's unlikely that a client
will ask two questions in the same second but it might be possible. If
that happens, and we're using a timestamp in seconds, then messages
could get lost between pages.