Add Anthropic dependency to Gemfile
We're going to port over from using the "aws-sdk-bedrockruntime" gem
to the "anthropic" gem for interacting with the Bedrock API.

We've run into a couple of issues using AWS's abstraction. The most recent
one is the lack of ability to cache individual system prompts.

This adds the new "anthropic" gem to the Gemfile.