Add :aws_credentials_stubbed tag to config
Overall i think it's better to use Webmock than just stubbing the interactions
via the client as it ensures that our requests adhere to the mock. And
we can easily check our prompts, user messages, tooling etc. align
with our expectations.

However, there's one issue with this appoarch. When we hit AWS Bedrock,
several requests realted to the credentials are made which we would have
to stub out as well.

After some research, it seems that the best way to avoid having to stub
out these requests is to temporarily stub the AWS credentials keys.

This commit adds a tag to the config that can be used to stub the access
and secret key.

These are reverted to the original values after the test run. This means
the only Webmock stubbing needed is is for the invoke requests.