Port the Claude::JailbreakChecker to the new SDK
This ports the Claude::JailbreakChecker to the new SDK. I'm going to keep
the pipeline steps in separate commits for ease of review.

It adds the first set of stubs for the new SDK. This creates a webmock
which expects the correct request and returns a stubbed response which
mirrors the real response. It uses the correct models etc.