riPort Claude::StructuredAnswerComposer to use the Anthropic SDK
This follows on from the previous commit and ports the
Claude::StructuredAnswerComposer to use the Anthropic SDK instead of the
Bedrock SDK.

It should be noted that as part of this commit i've broken up the system
prompt into two parts. One with static content and one with the dynamic
poriton (the contexts).

This allows us to make use of the Anthropic SDK's ability to cache individual
system promopts which was one of the motiviations for porting over to
their SDK in the first place.