Add conversation questions endpoint
This is the first step into adding pagination to the API when retrieving
questions for a conversation. Currently we just expose a single endpoint for conversations
which includes the conversation data, optional pending question and up
to 500 answered questions.

The idea is that we'll keep this existing endpoint (but limit it to a
more reasonable smaller number of answered questions) and then this new
questions endpoint will return just an array of answered questions.

If a conversation has only a few questions, the client only needs to
make a single API call to the existing endpoint. But if there are more,
they can hit this new endpoint to enumerate the questions page by page.

This commit adds the new endpoint at /api/v0/conversations/:id/questions
and returns a JSON array of the answered questions.

There's no actual pagination yet, that will follow.