Configure Rails/SaveBang to reduce inline rubocop disabling
This adds Rails/SaveBang AllowedReceivers option as a way for us to flag
attribute names that are unlikely to be ActiveRecord objects. This
prevents them from having Rails/SaveBang enforced on them.

Rails/SaveBang is a bit of a clunky rule as just about any attribute
could be an ActiveRecord object and the method names (save, update,
create) are super common.

We can only really customise this rule in individual applications as the
AllowedReceivers are domain specific.

While this only adds one usage, the expectation is that config and
messages will also be added to support: https://github.com/alphagov/govuk-chat/pull/304