Rename Device ID header
Renames the `HTTP_GOVUK_CHAT_CLIENT_DEVICE_ID` header to
`HTTP_GOVUK_CHAT_CLIENT_USER_ID` to make it a bit more clear as to its
purpose.

The "client device ID" part has confused people, as the app expects a user
to have same id on multiple devices.

This renames the header and all the supporting constant names, OpenAPI
spec, and so on.

It also bumps the API version as we're making a breaking change.