Inject answer_path into blueprints
At the moment we're using Rails.application.routes.url_helpers.api_v0_answer_question_path
in the question blueprint to generate the answer path for pending
questions. This is not ideal as it couples the blueprint to a specific
version of the API.

This updates the question blueprint to expect an answer_url option,
and passes the option to the blueprint from the conversations controller.