Add new ConversationQuestions model
The conversation questions API endpoint currently just returns an array
of answered questions.

We'll need to enhance this for pagination by adding other fields to this
response, so it makes sense to introduce a new class for this.

This does the groundwork for adding new fields; creating the class,
updating the OpenAPI spec and updating the tests. In a subsequent commit
we'll add the new fields.