Tidy up strict loading in specs
Because the answer factory doesn't set a `feedback` attribute, we often
need to do this trick in the specs to get around eager loading where we
have to re-query the database using `includes`.

If we just set a `nil` value in the factory, we can use the records
created using the factory and avoid a bunch of querying.