Handle conversation not found for no onboarding
As we no longer use the onboarding routes we don't want to redirect a
user to onboarding limitations if a conversation is not found. This
change changes the behaviour so that going to view conversation or
update conversation will behave as normal if a conversation cookie is
not there i.e. as if there is not an active conversation.

For places where we require a conversation - like checking an answer
status or clearing a conversation we now return a 404 response.

Returning a Rails 404 exception doesn't allow deleting the cookie, I
didn't think this was too big a concern as there is only a couple of
quite nested routes that would need this so I removed it from the test
assertion. However I've left a comment to indicate this issue to
hopefully save another developer from getting stuck if they do expect
cookies to be removed.

Doing this had allowed a decent simplification of the tests involved.