Admin area: filtering by end_user_id
This extends the admin panel question index view to allow filtering by
end_user_id.

It doesn't add an input field, but works by an admin viewing a question
and being able to click on the end_user link on the question show page,
which takes them to the questions index view filtered by that user.