Update get started messages for no onboarding
This replaces the get started messages with the new brief introduction
to GOV.UK Chat.

It also changes the behaviour of visiting the show page where it now
has the get started messages as new messages which are progressively
disclosed to a user if they don't have an active conversation and
rendered immediately if there is one.

This does introduce a minor annoyance that if you start a new chat
you'll get the progressively disclosed experience, even though you've
been through it before, but this doesn't seem too big a concern for our
web version and it's limited audience.