Limit access to conversation to end_user_id
If this `HTTP_GOVUK_CHAT_END_USER_ID` header is set then we'll only want
to find conversations where the `end_user_id` matches that header value.

This prevents a user from accessing anyone else's conversation.

This required a bit of a re-work of the specs for two reasons: firstly
because we want to test this behaviour on all the routes in the
controller except the POST :create, and secondly because I discovered a
small bug with the current implementation.

Before we had a `shared_examples` block with a bunch of routes passed in
as an argument. Doing it this way means you can't access anything in a
`let` block, so anywhere that a route was using a `conversation_id` param, we
had to use `SecureRandom.hex` as a fake value.

This mostly worked fine, except in the case of the spec that was testing
that we got a 404 if the conversation `source` was not `:api`. The test
was passing, but the 404 was being thrown because the conversation could
not be found (as it was trying to look up a conversation based on an ID
that doesn't exist).

It feels at this point that it's probably a good idea to create a shared
example block for each thing we want to test, and include it in each of
the controller action blocks we're testing. That way we get access to
the actual records created in the `let` blocks. It does of course have
the downside that it's more verbose.