Export end_user_id to BigQuery
This includes the conversation's end_user_id along with the question
when exporting to BigQuery.

In order to preserve anonymity, we hash the value before exporting.
This means all questions with the same end_user_id will have the same
hashed value in BigQuery, so you can still find all questions from the
same user, but you won't be able to find out the original client
end_user_id.