Remove onboarding code
This removes the routes, controllers, views and components related to
the onboarding part of GOV.UK Chat.

This has been removed to be replaced with a lighter onboarding process
that doesn't have multiple views.

I've left the JS code related to onboarding and that is something that
may be a little delicate to unpick, so it'll be it's own commmit.

I also noticed that there is a use of the onboarding form component
outside of the component [1], which will need resolving.

[1]: https://github.com/alphagov/govuk-chat/blob/a0b23c67f35467f0ff3d23e06dcceafa4bcc0cc2/app/views/conversations/clear.html.erb#L11