Fix tests for change of default answer strategy
This fixes tests that were broken by the change of the
Rails.configuration.answer_strategy value.

For most of these tests I have tried to make them agnostic to the answer
strategy so they wouldn't fail if we changed it back.

The exception to this is the js features system tests that feel like
they should have a real mocked answer strategy and that it should be the
default one.