Add Rake task for generating Titan embeddings
This task will query for 10,000 documents in the OpenSearch index where
the `titan_embedding` field is blank. It does so using scrolling[1] so as to
make the query more efficient.

It builds up an array of 10,000 document IDs and loops through that
array, kicking off a background job for each document. That background
job will query the OpenSearch index for the document, generate the
embedding and write it back to the index.

The reason we're querying again is that we don't want to pass the full
text content of the document into the background job, as that'll mean
it's stored in Redis. Some content can be quite long so we want to avoid
storing potentially large amounts of data in Redis.

[1]: https://1291fa08092a4a2dfd6d43f11d7bbcf3f22bf569.opensearch.org/docs/1.1/opensearch/rest-api/scroll/